Move all of the pumping related function from `BlanketLibrary` into the new generic Pumping class
