+++ process/models/blankets/hcpb.py
@@ -15,7 +15,10 @@ from process.models.engineering.ivc_functions import (
     calculate_pipe_bend_radius,
     pumping_powers_as_fractions,
 )
-from process.models.engineering.pumping import CoolantType
+from process.models.engineering.pumping import (
+    CoolantType,
+    pipe_hydraulic_diameter,
+)
 from process.models.power import PumpingPowerModelTypes
 from process.models.tfcoil.base import TFConductorModel
 
@@ -73,7 +76,12 @@ class CCFE_HCPB(OutboardBlanket, InboardBlanket):
             self.data.blanket.deg_blkt_inboard_poloidal_plasma / 360.0
         )
 
-        dia_blkt_channel = self.pipe_hydraulic_diameter(i_channel_shape=1)
+        dia_blkt_channel = pipe_hydraulic_diameter(
+            i_channel_shape=1,
+            radius_fw_channel=self.data.fwbs.radius_fw_channel,
+            a_bz_liq=self.data.fwbs.a_bz_liq,
+            b_bz_liq=self.data.fwbs.b_bz_liq,
+        )
         self.data.fwbs.radius_blkt_channel = dia_blkt_channel / 2
         (
             self.data.fwbs.radius_blkt_channel_90_bend,
