+++ tests/unit/models/blankets/test_blanket_library.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any, NamedTuple
+from typing import Any, NamedTuple
 
 import numpy as np
 import pytest
@@ -6,9 +6,6 @@ import pytest
 from process.data_structure.build_variables import InboardBlanketConfiguration
 from process.models.blankets.blanket_library import InboardBlanket
 from process.models.engineering.pumping import CoolantType
-
-if TYPE_CHECKING:
-    from process.models.engineering.pumping import CoolantFrictionLossParameters
 
 
 @pytest.fixture
@@ -19,6 +16,16 @@ def blanket_library(process_models):
     :rtype: process.blanket_library.BlanketLibrary
     """
     return process_models.blanket_library
+
+
+@pytest.fixture
+def pumping(process_models):
+    """Fixture to get the Physics instance from process_models.
+
+    :returns: initialised Physics object
+    :rtype: process.physics.Physics
+    """
+    return process_models.pumping
 
 
 class PrimaryCoolantPropertiesParam(NamedTuple):
@@ -343,50 +350,6 @@ def test_deltap_tot_outboard_blanket_breeder_liquid(monkeypatch, blanket_library
 
     dpres_total, _ = blanket_library.total_pressure_drop(False, **data)
     assert dpres_total == pytest.approx(56.95922064419226)
-
-
-def test_pumppower_primary_helium(monkeypatch, blanket_library):
-    monkeypatch.setattr(blanket_library.data.fwbs, "etaiso", 0.9)
-    monkeypatch.setattr(blanket_library.data.fwbs, "etaiso_liq", 0.85)
-
-    data = {
-        "i_liquid_breeder": 2,
-        "temp_coolant_pump_outlet": 570,
-        "temp_coolant_pump_inlet": 720,
-        "pres_coolant_pump_inlet": 1700000,
-        "dpres_coolant": 303517.3,
-        "mflow_coolant_total": 35677.7,
-        "i_coolant_type": 1,
-        "den_coolant": 9753.25,
-        "label": "Liquid Metal Breeder/Coolant",
-    }
-
-    assert (
-        pytest.approx(blanket_library.coolant_pumping_power(False, **data))
-        == 1.8251284651310427
-    )
-
-
-def test_pumppower_secondary_pb_li(monkeypatch, blanket_library):
-    monkeypatch.setattr(blanket_library.data.fwbs, "etaiso", 0.9)
-    monkeypatch.setattr(blanket_library.data.fwbs, "etaiso_liq", 0.85)
-
-    data = {
-        "i_liquid_breeder": 1,
-        "temp_coolant_pump_outlet": 573,
-        "temp_coolant_pump_inlet": 773,
-        "pres_coolant_pump_inlet": 8000000,
-        "dpres_coolant": 20088.23,
-        "mflow_coolant_total": 956.3,
-        "i_coolant_type": CoolantType.HELIUM,
-        "den_coolant": 5.64,
-        "label": "First Wall and Blanket",
-    }
-
-    assert (
-        pytest.approx(blanket_library.coolant_pumping_power(False, **data), rel=1e-4)
-        == 3.2374845432302464
-    )
 
 
 class ComponentHalfHeightParam(NamedTuple):
@@ -1025,109 +988,6 @@ def test_liquid_breeder_properties(
     )
 
 
-class CoolantFrictionLossParam(NamedTuple):
-    radius_channel: Any = None
-    radius_pipe_90_deg_bend: Any = None
-    radius_pipe_180_deg_bend: Any = None
-    a_bz_liq: Any = None
-    b_bz_liq: Any = None
-    roughness_channel: Any = None
-    i_ps: Any = None
-    n_pipe_90_deg_bends: Any = None
-    n_pipe_180_deg_bends: Any = None
-    len_pipe: Any = None
-    den_coolant: Any = None
-    visc_coolant: Any = None
-    vel_coolant: Any = None
-    label: Any = None
-    expected_pressure_drop_out: Any = None
-
-
-@pytest.mark.parametrize(
-    "coolantfrictionlossparam",
-    [
-        CoolantFrictionLossParam(
-            radius_channel=0.0060000000000000001,
-            radius_pipe_90_deg_bend=0.018,
-            radius_pipe_180_deg_bend=0.09,
-            a_bz_liq=0.20000000000000001,
-            b_bz_liq=0.20000000000000001,
-            roughness_channel=9.9999999999999995e-07,
-            i_ps=1,
-            n_pipe_90_deg_bends=2,
-            n_pipe_180_deg_bends=0,
-            len_pipe=4,
-            den_coolant=10.405276820718059,
-            visc_coolant=3.604452999475736e-05,
-            vel_coolant=32.753134225223164,
-            label="Inboard first wall",
-            expected_pressure_drop_out=36213.58989742931,
-        ),
-        CoolantFrictionLossParam(
-            radius_channel=1.0,
-            radius_pipe_90_deg_bend=1.0,
-            radius_pipe_180_deg_bend=1.0,
-            a_bz_liq=1.0,
-            b_bz_liq=1.0,
-            roughness_channel=1e-6,
-            i_ps=2,
-            n_pipe_90_deg_bends=1.0,
-            n_pipe_180_deg_bends=1.0,
-            len_pipe=1.0,
-            den_coolant=1.0,
-            visc_coolant=1.0,
-            vel_coolant=1.0,
-            label="label",
-            expected_pressure_drop_out=1.4325633520224854,
-        ),
-    ],
-)
-def test_coolant_friction_loss(coolantfrictionlossparam, monkeypatch, blanket_library):
-    """
-    Automatically generated Unit Test for pressure_drop.
-
-    This test was generated using data from
-    blanket_files/large_tokamak_primary_pumping2.IN.DAT.
-
-    Parameters
-    ----------
-    coolantfrictionlossparam : CoolantFrictionLossParam
-        the data used to mock and assert in this test.
-    monkeypatch : _pytest.monkeypatch.monkeypatch
-        pytest fixture used to mock module/class variables
-    blanket_library : BlanketLibrary
-        the blanket library instance used in this test.
-
-    """
-
-    monkeypatch.setattr(
-        blanket_library.data.fwbs, "a_bz_liq", coolantfrictionlossparam.a_bz_liq
-    )
-    monkeypatch.setattr(
-        blanket_library.data.fwbs, "b_bz_liq", coolantfrictionlossparam.b_bz_liq
-    )
-
-    pressure_params: CoolantFrictionLossParameters = (
-        blanket_library.coolant_friction_pressure_drop(
-            i_ps=coolantfrictionlossparam.i_ps,
-            radius_pipe_90_deg_bend=(coolantfrictionlossparam.radius_pipe_90_deg_bend),
-            radius_pipe_180_deg_bend=(coolantfrictionlossparam.radius_pipe_180_deg_bend),
-            n_pipe_90_deg_bends=coolantfrictionlossparam.n_pipe_90_deg_bends,
-            n_pipe_180_deg_bends=coolantfrictionlossparam.n_pipe_180_deg_bends,
-            len_pipe=coolantfrictionlossparam.len_pipe,
-            den_coolant=coolantfrictionlossparam.den_coolant,
-            visc_coolant=coolantfrictionlossparam.visc_coolant,
-            vel_coolant=coolantfrictionlossparam.vel_coolant,
-            roughness_channel=coolantfrictionlossparam.roughness_channel,
-            radius_channel=coolantfrictionlossparam.radius_channel,
-        )
-    )
-
-    assert pressure_params.dpres_total == pytest.approx(
-        coolantfrictionlossparam.expected_pressure_drop_out
-    )
-
-
 class LiquidBreederPressureDropMhdParam(NamedTuple):
     i_blkt_liquid_breeder_channel_type: Any = None
     a_bz_liq: Any = None
@@ -1590,22 +1450,6 @@ def test_calculate_elliptical_blkt_volumes(
     assert vol_blkt_total == pytest.approx(
         calculateellipticalblktvolumesparam.expected_vol_blkt_total
     )
-
-
-def test_hydraulic_diameter(monkeypatch, blanket_library):
-    """
-    Test for hydraulic_diameter function.
-    """
-    # Set var values
-    monkeypatch.setattr(blanket_library.data.fwbs, "radius_fw_channel", 1.0)
-    monkeypatch.setattr(blanket_library.data.fwbs, "a_bz_liq", 1.0)
-    monkeypatch.setattr(blanket_library.data.fwbs, "b_bz_liq", 1.0)
-
-    # hydraulic_diameter input = i_channel_shape: 1 = circle, 2 = rectangle
-    # 2.0D0*radius_fw_channel
-    assert blanket_library.pipe_hydraulic_diameter(1) == pytest.approx(2.0)
-    # 2*a_bz_liq*b_bz_liq/(a_bz_liq+b_bz_liq)
-    assert blanket_library.pipe_hydraulic_diameter(2) == pytest.approx(1.0)
 
 
 def test_flow_velocity(monkeypatch, blanket_library):
